% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edmdata-package.R
\docType{package}
\name{edmdata-package}
\alias{edmdata}
\alias{edmdata-package}
\title{edmdata: Data Sets for Psychometric Modeling}
\description{
Collection of data sets from assessments that can be used to 
    evaluate psychometric models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tmsalab.github.io/edmdata}
  \item \url{https://github.com/tmsalab/edmdata}
  \item Report bugs at \url{https://github.com/tmsalab/edmdata/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID})

Authors:
\itemize{
  \item Steven Andrew Culpepper \email{sculpepp@illinois.edu} (\href{https://orcid.org/0000-0003-4226-6176}{ORCID})
  \item Jeffrey Alan Douglas \email{jeffdoug@illinois.edu}
}

}
\keyword{internal}
