% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcp-penn-matrix.R
\docType{data}
\name{items_hcp_penn_matrix}
\alias{items_hcp_penn_matrix}
\alias{items_hcp_penn_matrix_missing}
\title{Human Connectome Project's Penn Progressive Matrices Fluid Intelligence Assessment}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1201 rows and 24 columns.

An object of class \code{matrix} (inherits from \code{array}) with 1201 rows and 24 columns.
}
\usage{
items_hcp_penn_matrix

items_hcp_penn_matrix_missing
}
\description{
Trial data from Form A of an abbreviated version of the
Raven’s Progressive Matrices developed by Gur and colleagues (Bilker et al. 2012).
Participants are presented with patterns made up of 2x2, 3x3 or 1x5 arrangements of squares, with one of the squares missing.
The participant must pick one of five response choices that best fits the missing square on the pattern.
The task has 24 items and 3 bonus items, arranged in order of increasing difficulty.
However, the task discontinues if the participant makes 5 incorrect responses in a row.
}
\details{
The subjects answered the following assessment items:
\itemize{
\item \code{item01}
\item \code{item02}
\item \code{item03}
\item \code{item04}
\item \code{item05}
\item \code{item06}
\item \code{item07}
\item \code{item08}
\item \code{item09}
\item \code{item11}
\item \code{item12}
\item \code{item13}
\item \code{item14}
\item \code{item15}
\item \code{item16}
\item \code{item17}
\item \code{item18}
\item \code{item19}
\item \code{item20}
\item \code{item21}
\item \code{item22}
\item \code{item23}
\item \code{item24}
}
}
\section{Data Cleaning}{

The \code{items_hcp_penn} data set includes \eqn{N = 1,201} observations.
However, the observation count was calculated after removing participants
that did not respond to any questions (Subject ID's:
117728, 137431, 145531, 236130, and 614439). From there, we generated the
\code{items_hcp_penn_missing} data set that retains a missing data structure
after the participant makes 5 incorrect responses in a row.
}

\references{
\strong{Data originated from:}

Data were provided in part by the Human Connectome Project,
WU-Minn Consortium (Principal Investigators: David Van Essen and Kamil Ugurbil; 1U54MH091657)
funded by the 16 NIH Institutes and Centers that support the NIH Blueprint for Neuroscience Research;
and by the McDonnell Center for Systems Neuroscience at Washington University.

\strong{Data used in:}
\itemize{
\item TBA
}
}
\keyword{datasets}
