% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability-data.R
\docType{data}
\name{items_probability_part_one_full}
\alias{items_probability_part_one_full}
\alias{items_probability_part_one_reduced}
\title{Elementary Probability Theory Assessment Item Responses}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 504 rows and 12 columns.

An object of class \code{matrix} (inherits from \code{array}) with 431 rows and 12 columns.
}
\usage{
items_probability_part_one_full

items_probability_part_one_reduced
}
\description{
Elementary Probability Theory Assessment Item Responses
}
\details{
Questions wording and answers are from the \code{pks} package documentation.

Items with their desired responses bolded:
\itemize{
\item \code{b101}:  A box contains 30 marbles in the following colors:
8 red, 10 black, 12 yellow.  What is the probability that a randomly
drawn marble is yellow?  \strong{(0.40)}
\item \code{b102}:  A bag contains 5-cent, 10-cent, and 20-cent coins.
The probability of drawing a 5-cent coin is 0.35, that of drawing a
10-cent coin is 0.25, and that of drawing a 20-cent coin is 0.40.  What
is the probability that the coin randomly drawn is not a 5-cent
coin?  \strong{(0.65)}
\item \code{b103}:  A bag contains 5-cent, 10-cent, and 20-cent coins.
The probability of drawing a 5-cent coin is 0.20, that of drawing a
10-cent coin is 0.45, and that of drawing a 20-cent coin is 0.35.  What
is the probability that the coin randomly drawn is a 5-cent coin or
a 20-cent coin?  \strong{(0.55)}
\item \code{b104}:  In a school, 40\\% of the pupils are boys and 80\\% of
the pupils are right-handed.  Suppose that gender and handedness are
independent.  What is the probability of randomly selecting a
right-handed boy?  \strong{(0.32)}
\item \code{b105}:  Given a standard deck containing 32 different cards,
what is the probability of not drawing a heart?  \strong{(0.75)}
\item \code{b106}:  A box contains 20 marbles in the following colors:
4 white, 14 green, 2 red.  What is the probability that a randomly
drawn marble is not white?  \strong{(0.80)}
\item \code{b107}:  A box contains 10 marbles in the following colors:
2 yellow, 5 blue, 3 red.  What is the probability that a randomly
drawn marble is yellow or blue?  \strong{(0.70)}
\item \code{b108}:  What is the probability of obtaining an even number
by throwing a dice?  \strong{(0.50)}
\item \code{b109}:  Given a standard deck containing 32 different cards,
what is the probability of drawing a 4 in a black suit?
\strong{(Responses that round to 0.06 were considered correct.)}
\item \code{b110}:  A box contains marbles that are red or yellow, small
or large. The probability of drawing a red marble is 0.70 \strong{(lab: 0.30)},
the probability of drawing a small marble is 0.40.  Suppose that the
color of the marbles is independent of their size.  What is the
probability of randomly drawing a small marble that is not red?  \strong{(0.12, lab: 0.28)}
\item \code{b111}:  In a garage there are 50 cars.  20 are black and 10 are
diesel powered.  Suppose that the color of the cars is independent of
the kind of fuel.  What is the probability that a randomly selected car
is not black and it is diesel powered?  \strong{(0.12)}
\item \code{b112}:  A box contains 20 marbles.  10 marbles are red, 6 are
yellow and 4 are black.  12 marbles are small and 8 are large.  Suppose
that the color of the marbles is independent of their size.  What is the
probability of randomly drawing a small marble that is yellow or
red?  \strong{(0.48)}
}
}
\section{Data Cleaning}{

The \code{items_probability_part_one_full} data set includes \eqn{N = 504} observations
used in Chen et al. (2021). However, the observations contained
missing trial data as \code{NA}. Excluding the missing response values from part
one respondents yields \eqn{N = 431}. The data set with reduced items is given
in \code{items_probability_part_one_reduced}.
}

\references{
\strong{Data originated from:}
\itemize{
\item Heller, J., & Wickelmaier, F. (2013).
Minimum discrepancy estimation in probabilistic knowledge structures.
Electronic Notes in Discrete Mathematics, 42, 49–56.
\doi{10.1016/j.endm.2013.05.145}
}

\strong{Data used in:}
\itemize{
\item Chen, Yinghan, Liu, Y., Culpepper, S. A., & Chen, Y. (2021).
Inferring the number of attributes for the exploratory DINA model.
Psychometrika, 86(1), 30–64. \doi{10.1007/s11336-021-09750-9}
}
}
\keyword{datasets}
