% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecpe-data.R
\docType{data}
\name{items_ecpe}
\alias{items_ecpe}
\title{Examination for the Certificate of Proficiency in English (ECPE) Item Responses}
\format{
A \code{matrix} consisting of 2922 rows and 28 columns. Thus, there are
\eqn{N = 2922} subject responses to \eqn{J = 28} items.

The subjects answered the following assessment items:
\itemize{
\item \code{item01}
\item \code{item02}
\item \code{item03}
\item \code{item04}
\item \code{item05}
\item \code{item06}
\item \code{item07}
\item \code{item08}
\item \code{item09}
\item \code{item11}
\item \code{item12}
\item \code{item13}
\item \code{item14}
\item \code{item15}
\item \code{item16}
\item \code{item17}
\item \code{item18}
\item \code{item19}
\item \code{item20}
\item \code{item21}
\item \code{item22}
\item \code{item23}
\item \code{item24}
\item \code{item25}
\item \code{item26}
\item \code{item27}
\item \code{item28}
}
}
\usage{
items_ecpe
}
\description{
Examination for the Certificate of Proficiency in English (ECPE) Item Responses
}
\references{
Tatsuoka, C. (2002), Data analytic methods for latent partially ordered
classification models. Journal of the Royal Statistical Society: Series C
(Applied Statistics), 51: 337–350. \url{doi:10.1111/1467-9876.00272}

Tatsuoka, K. K. (1984), Analysis of errors in fraction addition and subtraction
problems (Final Report for Grant No. NIE-G-81-0002). Urbana: University of
Illinois, Computer-Based Education Research Laboratory (CERL).
}
\keyword{datasets}
