% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-reasoning-data.R
\docType{data}
\name{items_matrix_reasoning}
\alias{items_matrix_reasoning}
\title{Experimental Matrix Reasoning Test Item Responses}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 400 rows and 25 columns.
}
\usage{
items_matrix_reasoning
}
\description{
Experimental Matrix Reasoning Test Item Responses
}
\details{
Items included:
\itemize{
\item \code{Q1}
\item \code{Q2}
\item \code{Q3}
\item \code{Q4}
\item \code{Q5}
\item \code{Q6}
\item \code{Q7}
\item \code{Q8}
\item \code{Q9}
\item \code{Q10}
\item \code{Q11}
\item \code{Q12}
\item \code{Q13}
\item \code{Q14}
\item \code{Q15}
\item \code{Q16}
\item \code{Q17}
\item \code{Q18}
\item \code{Q19}
\item \code{Q20}
\item \code{Q21}
\item \code{Q22}
\item \code{Q23}
\item \code{Q24}
\item \code{Q25}
}
}
\section{Answer coding}{

The subjects answered a set of assessment items seeking to determine
their matrix reasoning abilities. Subjects that answered with a value between
0 to 7 were marked as incorrect. Subjects who answered a question with 10
selected the correct answer and, thus, were marked as correct.
}

\section{Notes}{


From the OpenPsychometrics' code book that accompanied the data, they noted:
\enumerate{
\item The possible answers were presented in two rows of four with
a random order for each participant.
\item The collection of this data was of mediocre quality.
}
}

\references{
\strong{Data originated from:}
\itemize{
\item OpenPsychometrics. (2012). \emph{Experimental matrix reasoning iq test}.
\url{https://openpsychometrics.org/_rawdata/IQ1.zip}
}

\strong{Data used in:}
\itemize{
\item Chen, Y., Culpepper, S. A., & Liang, F. (2020). A sparse latent class model
for cognitive diagnosis. \emph{Psychometrika}, 1–33.
\doi{10.1007/s11336-019-09693-2}
}
}
\keyword{datasets}
