% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narcissistic-personality-inventory-data.R
\docType{data}
\name{items_narcissistic_personality_inventory}
\alias{items_narcissistic_personality_inventory}
\title{Narcissistic Personality Inventory Item Responses}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 11243 rows and 40 columns.
}
\usage{
items_narcissistic_personality_inventory
}
\description{
Narcissistic Personality Inventory Item Responses
}
\details{
Items with their desired option response bolded:
\itemize{
\item \code{Q1}
\itemize{
\item \strong{Option 1}: I have a natural talent for influencing people
\item Option 2: I am not good at influencing people.
}
\item \code{Q2}
\itemize{
\item \strong{Option 1}: Modesty doesn't become me
\item Option 2: I am essentially a modest person.
}
\item \code{Q3}
\itemize{
\item \strong{Option 1}: I would do almost anything on a dare
\item Option 2: I tend to be a fairly cautious person.
}
\item \code{Q4}
\itemize{
\item Option 1: When people compliment me I sometimes get embarrassed
\item \strong{Option 2}: I know that I am good because everybody keeps telling me so.
}
\item \code{Q5}
\itemize{
\item Option 1: The thought of ruling the world frightens the hell out of me
\item \strong{Option 2}: If I ruled the world it would be a better place.
}
\item \code{Q6}
\itemize{
\item \strong{Option 1}: I can usually talk my way out of anything
\item Option 2: I try to accept the consequences of my behavior.
}
\item \code{Q7}
\itemize{
\item Option 1: I prefer to blend in with the crowd
\item \strong{Option 2}: I like to be the center of attention.
}
\item \code{Q8}
\itemize{
\item \strong{Option 1}: I will be a success
\item Option 2: I am not too concerned about success.
}
\item \code{Q9}
\itemize{
\item Option 1: I am no better or worse than most people
\item \strong{Option 2}: I think I am a special person.
}
\item \code{Q10}
\itemize{
\item Option 1: I am not sure if I would make a good leader
\item \strong{Option 2}: I see myself as a good leader.
}
\item \code{Q11}
\itemize{
\item \strong{Option 1}: I am assertive
\item Option 2: I wish I were more assertive.
}
\item \code{Q12}
\itemize{
\item \strong{Option 1}: I like to have authority over other people
\item Option 2: I don't mind following orders.
}
\item \code{Q13}
\itemize{
\item \strong{Option 1}: I find it easy to manipulate people
\item Option 2: I don't like it when I find myself manipulating people.
}
\item \code{Q14}
\itemize{
\item \strong{Option 1}: I insist upon getting the respect that is due me
\item Option 2: I usually get the respect that I deserve.
}
\item \code{Q15}
\itemize{
\item Option 1: I don't particularly like to show off my body
\item \strong{Option 2}: I like to show off my body.
}
\item \code{Q16}
\itemize{
\item \strong{Option 1}: I can read people like a book
\item Option 2: People are sometimes hard to understand.
}
\item \code{Q17}
\itemize{
\item Option 1: If I feel competent I am willing to take responsibility for making decisions
\item \strong{Option 2}: I like to take responsibility for making decisions.
}
\item \code{Q18}
\itemize{
\item Option 1: I just want to be reasonably happy
\item \strong{Option 2}: I want to amount to something in the eyes of the world.
}
\item \code{Q19}
\itemize{
\item Option 1: My body is nothing special
\item \strong{Option 2}: I like to look at my body.
}
\item \code{Q20}
\itemize{
\item Option 1: I try not to be a show off
\item \strong{Option 2}: I will usually show off if I get the chance.
}
\item \code{Q21}
\itemize{
\item \strong{Option 1}: I always know what I am doing
\item Option 2: Sometimes I am not sure of what I am doing.
}
\item \code{Q22}
\itemize{
\item Option 1: I sometimes depend on people to get things done
\item \strong{Option 2}: I rarely depend on anyone else to get things done.
}
\item \code{Q23}
\itemize{
\item Option 1: Sometimes I tell good stories
\item \strong{Option 2}: Everybody likes to hear my stories.
}
\item \code{Q24}
\itemize{
\item \strong{Option 1}: I expect a great deal from other people
\item Option 2: I like to do things for other people.
}
\item \code{Q25}
\itemize{
\item \strong{Option 1}: I will never be satisfied until I get all that I deserve
\item Option 2: I take my satisfactions as they come.
}
\item \code{Q26}
\itemize{
\item Option 1: Compliments embarrass me
\item \strong{Option 2}: I like to be complimented.
}
\item \code{Q27}
\itemize{
\item \strong{Option 1}: I have a strong will to power
\item Option 2: Power for its own sake doesn't interest me.
}
\item \code{Q28}
\itemize{
\item Option 1: I don't care about new fads and fashions
\item \strong{Option 2}: I like to start new fads and fashions.
}
\item \code{Q29}
\itemize{
\item \strong{Option 1}: I like to look at myself in the mirror
\item Option 2: I am not particularly interested in looking at myself in the mirror.
}
\item \code{Q30}
\itemize{
\item \strong{Option 1}: I really like to be the center of attention
\item Option 2: It makes me uncomfortable to be the center of attention.
}
\item \code{Q31}
\itemize{
\item \strong{Option 1}: I can live my life in any way I want to
\item Option 2: People can't always live their lives in terms of what they want.
}
\item \code{Q32}
\itemize{
\item Option 1: Being an authority doesn't mean that much to me
\item \strong{Option 2}: People always seem to recognize my authority.
}
\item \code{Q33}
\itemize{
\item \strong{Option 1}: I would prefer to be a leader
\item Option 2: It makes little difference to me whether I am a leader or not.
}
\item \code{Q34}
\itemize{
\item \strong{Option 1}: I am going to be a great person
\item Option 2: I hope I am going to be successful.
}
\item \code{Q35}
\itemize{
\item Option 1: People sometimes believe what I tell them
\item \strong{Option 2}: I can make anybody believe anything I want them to.
}
\item \code{Q36}
\itemize{
\item \strong{Option 1}: I am a born leader
\item Option 2: Leadership is a quality that takes a long time to develop.
}
\item \code{Q37}
\itemize{
\item \strong{Option 1}: I wish somebody would someday write my biography
\item Option 2: I don't like people to pry into my life for any reason.
}
\item \code{Q38}
\itemize{
\item \strong{Option 1}: I get upset when people don't notice how I look when I go out in public
\item Option 2: I don't mind blending into the crowd when I go out in public.
}
\item \code{Q39}
\itemize{
\item \strong{Option 1}: I am more capable than other people
\item Option 2: There is a lot that I can learn from other people.
}
\item \code{Q40}
\itemize{
\item Option 1: I am much like everybody else
\item \strong{Option 2}: I am an extraordinary person.
}
}
}
\section{Data pre-processing}{

We have applied list-wise deletion during pre-processing to remove any
observations with missing values from the data set.
}

\section{Answer coding}{

The subjects answered a set of assessment items seeking to determine
the level of anxiety. Answers given in \strong{bold} represent the desired
response. If a subject matched this response, they were given a 1 inside
of the item matrix, otherwise they received a zero.
}

\references{
\strong{Assessment Design:}
\itemize{
\item Raskin, R., & Terry, H. (1988). A principal-components analysis of the
narcissistic personality inventory and further evidence of its construct
validity. \emph{Journal of Personality and Social Psychology}, \emph{54}(5), 890.
\url{https://doi.org/10.1037/0022-3514.54.5.890}
}

\strong{Data originated from:}
\itemize{
\item OpenPsychometrics. 2013. \emph{Narcissistic Personality Inventory}.
\url{https://openpsychometrics.org/_rawdata/NPI.zip}.
}

\strong{Data used in:}
\itemize{
\item TBA
}
}
\keyword{datasets}
