% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisa12-us-vignette-data.R
\docType{data}
\name{items_ordered_pisa12_us_vignette}
\alias{items_ordered_pisa12_us_vignette}
\title{Programme for International Student Assessment (PISA) 2012
U.S. Student Questionnaire Problem-Solving Vignettes}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3075 rows and 12 columns.
}
\usage{
items_ordered_pisa12_us_vignette
}
\description{
Programme for International Student Assessment (PISA) 2012
U.S. Student Questionnaire Problem-Solving Vignettes
}
\details{
Obtained from the PISA 2012 questionnaire responses given to 15-year olds,
we selected 12 items under three vignettes and four item stems to measure
problem-solving strategies applied by students.
\itemize{
\item Vignette \strong{(a)}:
\itemize{
\item Suppose that you have been sending text messages from your mobile phone
for several weeks.Today, however, you can’t send text messages.
You want to try to solve the problem.
\itemize{
\item \strong{ST96Q01}: I press every button possible to find out what is wrong.
\item \strong{ST96Q02}: I think about what might have caused the problem and what I can do to solve it.
\item \strong{ST96Q03}: I read the manual.
\item \strong{ST96Q05}: I ask a friend for help.
}
}
\item Vignette \strong{(b)}:
\itemize{
\item Suppose that you are planning a trip to the zoo with your brother.
You don’t know which route to take to get there.
\itemize{
\item \strong{ST101Q01}: I read the zoo brochure to see if it says how to get there.
\item \strong{ST101Q02}: I study a map and work out the best route.
\item \strong{ST101Q03}: I leave it to my brother to worry about how to get there.
\item \strong{ST101Q05}: I know roughly where it is, so I suggest we just start driving.
}
}
\item Vignette \strong{(c)}:
\itemize{
\item Suppose that you arrive at the train station.
There is a ticket machine that you have never used before.
You want to buy a ticket.
\itemize{
\item \strong{ST104Q01}: I check how similar it is to other ticket machines I have used.
\item \strong{ST104Q04}: I try out all the buttons to see what happens.
\item \strong{ST104Q05}: I ask someone for help.
\item \strong{ST104Q06}: I try to find a ticket office at the station to buy a ticket.
}
}
}
}
\section{Answer Coding}{

All items were reverse-coded such that responses categories begin with
negative responses at 0 and go to 3 for a positive responses.
\itemize{
\item \strong{0}: I would definitely not do this.
\item \strong{1}: I would probably not do this.
\item \strong{2}: I would probably do this.
\item \strong{3}: I would definitely do this.
}
}

\references{
\strong{Assessment Design:}
\itemize{
\item PISA Test Overview: \url{https://www.oecd.org/pisa/aboutpisa.html}
\item PISA 2012 Mathematics Items:
\url{https://www.oecd.org/pisa/test/PISA\%202012\%20items\%20for\%20release_ENGLISH.pdf}
}

\strong{Data used in:}
\itemize{
\item Culpepper, S. A., & Balamuta, J. J. (2021).
Inferring latent structure in polytomous data with a higher-order diagnostic model.
Multivariate Behavioral Research, 1–19. \doi{10.1080/00273171.2021.1985949}
}
}
\keyword{datasets}
