% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordered-pswc-data.R
\docType{data}
\name{items_ordered_pswc_hw}
\alias{items_ordered_pswc_hw}
\title{Calculus-based probability and statistics course homework problems}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 288 rows and 29 columns.
}
\usage{
items_ordered_pswc_hw
}
\description{
Calculus-based probability and statistics course homework problems
}
\details{
Data set contains the subject's responses to calculus-based probability and
statistics course homework problems.
\itemize{
\item \code{Item01}: Subject's Response to Item 1.
\item \code{Item02}: Subject's Response to Item 2.
\item \code{Item03}: Subject's Response to Item 3.
\item \code{Item05}: Subject's Response to Item 5.
\item \code{Item06}: Subject's Response to Item 6.
\item \code{Item07}: Subject's Response to Item 7.
\item \code{Item08}: Subject's Response to Item 8.
\item \code{Item09}: Subject's Response to Item 9.
\item \code{Item10}: Subject's Response to Item 10.
\item \code{Item11}: Subject's Response to Item 11.
\item \code{Item12}: Subject's Response to Item 12.
\item \code{Item13}: Subject's Response to Item 13.
\item \code{Item14}: Subject's Response to Item 14.
\item \code{Item15}: Subject's Response to Item 15.
\item \code{Item16}: Subject's Response to Item 16.
\item \code{Item17}: Subject's Response to Item 17.
\item \code{Item18}: Subject's Response to Item 18.
\item \code{Item19}: Subject's Response to Item 19.
\item \code{Item21}: Subject's Response to Item 21.
\item \code{Item22}: Subject's Response to Item 22.
\item \code{Item23}: Subject's Response to Item 23.
\item \code{Item24}: Subject's Response to Item 24.
\item \code{Item25}: Subject's Response to Item 25.
\item \code{Item26}: Subject's Response to Item 26.
\item \code{Item27}: Subject's Response to Item 27.
\item \code{Item29}: Subject's Response to Item 29.
\item \code{Item30}: Subject's Response to Item 30.
\item \code{Item32}: Subject's Response to Item 32.
\item \code{Item34}: Subject's Response to Item 34.
}
}
\references{
\strong{Data used in:}
\itemize{
\item Jimenez, A., Balamuta, J. J., & Culpepper, S. A. (2023).
A sequential exploratory diagnostic model using a Pólya-gamma data
augmentation strategy. \emph{British Journal of Mathematical and Statistical Psychology},
\emph{76}, 513–538. \doi{10.1111/bmsp.12307}
}

\strong{Data originated from:}
\itemize{
\item Culpepper, S. A. (2014).
If at first you don’t succeed, try, try again: Applications of sequential
IRT models to cognitive assessments. \emph{Applied Psychological Measurement},
38(8), 632–644. \doi{10.1177/0146621614536464}
}
}
\keyword{datasets}
