% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timss15-background-data.R
\docType{data}
\name{items_ordered_timss15_background}
\alias{items_ordered_timss15_background}
\title{Trends in International Mathematics and Science Study 2015 (TIMSS)
Grade 8 Student Background Survey Item Responses}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 9672 rows and 16 columns.
}
\usage{
items_ordered_timss15_background
}
\description{
Trends in International Mathematics and Science Study 2015 (TIMSS)
Grade 8 Student Background Survey Item Responses
}
\details{
From the school background data file, we selected 16 items under two distinct
item stems that sought to measure students’ school experiences using a
four-level likert-scale.
\itemize{
\item What do you think about your school? Tell me howmuch you agree with these statements:
\itemize{
\item \strong{"BSBG15A"}: I like being in school
\item \strong{"BSBG15B"}: I feel safe when I am in school
\item \strong{"BSBG15C"}: I feel like I belong at this school
\item \strong{"BSBG15D"}: I like to see my classmates at school.
\item \strong{"BSBG15E"}: Teachers at my school are fair to me.
\item \strong{"BSBG15F"}: I am proud to go to this school.
\item \strong{"BSBG15G"}: I learn a lot in school.
}
\item During this year, how often have the following things happened to you at school?
\itemize{
\item \strong{"BSBG16A"}: Made fun of me or called me names.
\item \strong{"BSBG16B"}: Left me out of their games or activities.
\item \strong{"BSBG16C"}: Spread lies about me.
\item \strong{"BSBG16D"}: Stole something from me.
\item \strong{"BSBG16E"}: Hit or hurt me (e.g. shoving, hitting, kicking).
\item \strong{"BSBG16F"}: Made me do things I didn’t want to.
\item \strong{"BSBG16G"}: Shared embarrassing information about me.
\item \strong{"BSBG16H"}: Posted embarrassing things about me online.
\item \strong{"BSBG16I"}: Threatened me.
}
}
}
\section{Answer Coding}{

Items 8 through 16 were reverse-coded.
}

\references{
\strong{Data originated from:}
\itemize{
\item Trends in International Mathematics and Science Study 2015 (TIMSS) Grade 8
Student Background Survey Item Responses (Mullis et al., 2016).
}

\strong{Data used in:}
\itemize{
\item TBA
}
}
\keyword{datasets}
