% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisa12-us-math.R
\docType{data}
\name{items_pisa12_us_math}
\alias{items_pisa12_us_math}
\title{Programme for International Student Assessment (PISA) 2012 US Math Assessment}
\format{
An object of class \code{data.frame} with 4978 rows and 76 columns.
}
\usage{
items_pisa12_us_math
}
\description{
Programme for International Student Assessment (PISA) 2012 US Math Assessment
}
\details{
Items alongside of the problem name:
}
\section{Answer coding}{

The subjects answered a set of assessment items seeking to determine
their knowledge of mathematics. The original PISA 2012 data set included
variables that had problems with the following labels:
\itemize{
\item One level:
\itemize{
\item "N/A"
}
\item Four Levels:
\itemize{
\item "Score 0"
\item "Score 1"
\item "Score 7: N/A"
\item "Score 8: Not Attempted"
}
\item Five Levels:
\itemize{
\item "Score 0"
\item "Score 1"
\item "Score 2"
\item "Score 7: N/A"
\item "Score 8: Not Attempted"
}
}

Note, this data set includes items with four levels only. Further,
items are coded with \verb{Score 0} as \code{0}, \verb{Score 1} as \code{1},
\verb{Score 7: N/A} as \code{NA}, and \verb{Score 8: Not Attempted} as 0.
}

\references{
\strong{Assessment Design:}
\itemize{
\item PISA Test Overview: https://www.oecd.org/pisa/aboutpisa/
\item PISA 2012 Mathematics Items:
\url{https://www.oecd.org/pisa/test/PISA\%202012\%20items\%20for\%20release_ENGLISH.pdf}
}

\strong{Data originated from:}
\itemize{
\item Paul Bailey, Ahmad Emad, Huade Huo, Michael Lee, Yuqi Liao, Alex Lishinski, Trang Nguyen,
Qingshu Xie, Jiao Yu, Ting Zhang, Eric Buehler, Jeppe Bundsgaard,
Ren C'deBaca and Anders Astrup Christensen (2021).
EdSurvey: Analysis of NCES Education Survey
and Assessment Data. R package version 2.7.0.
\url{https://CRAN.R-project.org/package=EdSurvey}
}

\strong{Data used in:}
\itemize{
\item TBA
}
}
\keyword{datasets}
