% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revised-psvtr-data.R
\docType{data}
\name{items_revised_psvtr}
\alias{items_revised_psvtr}
\title{Revised PSVT:R Item Responses}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 516 rows and 30 columns.
}
\usage{
items_revised_psvtr
}
\description{
Revised PSVT:R Item Responses
}
\details{
Data set contains the subject's responses to Revised PSVT:R items.
Correct answers are denoted by 1 and incorrect answers are denoted by 0.
\itemize{
\item \code{Item01}: Subject's Response to Item 1.
\item \code{Item02}: Subject's Response to Item 2.
\item \code{Item03}: Subject's Response to Item 3.
\item \code{Item04}: Subject's Response to Item 4.
\item \code{Item05}: Subject's Response to Item 5.
\item \code{Item06}: Subject's Response to Item 6.
\item \code{Item07}: Subject's Response to Item 7.
\item \code{Item08}: Subject's Response to Item 8.
\item \code{Item09}: Subject's Response to Item 9.
\item \code{Item10}: Subject's Response to Item 10.
\item \code{Item11}: Subject's Response to Item 11.
\item \code{Item12}: Subject's Response to Item 12.
\item \code{Item13}: Subject's Response to Item 13.
\item \code{Item14}: Subject's Response to Item 14.
\item \code{Item15}: Subject's Response to Item 15.
\item \code{Item16}: Subject's Response to Item 16.
\item \code{Item17}: Subject's Response to Item 17.
\item \code{Item18}: Subject's Response to Item 18.
\item \code{Item19}: Subject's Response to Item 19.
\item \code{Item20}: Subject's Response to Item 20.
\item \code{Item21}: Subject's Response to Item 21.
\item \code{Item22}: Subject's Response to Item 22.
\item \code{Item23}: Subject's Response to Item 23.
\item \code{Item24}: Subject's Response to Item 24.
\item \code{Item25}: Subject's Response to Item 25.
\item \code{Item26}: Subject's Response to Item 26.
\item \code{Item27}: Subject's Response to Item 27.
\item \code{Item28}: Subject's Response to Item 28.
\item \code{Item29}: Subject's Response to Item 29.
\item \code{Item30}: Subject's Response to Item 30.
}
}
\references{
\strong{Assessment Design:}
\itemize{
\item Yoon, S. Y. (2011). \emph{Psychometric properties of the revised purdue
spatial visualization tests: Visualization of rotations (the revised
psvt: R)}. Purdue University.
}

\strong{Data originated from:}
\itemize{
\item Culpepper, S. A., & Balamuta, J. J. (2017). A Hierarchical Model for
Accuracy and Choice on Standardized Tests. \emph{Psychometrika}, \emph{82}(3),
820–845. \url{https://doi.org/10.1007/s11336-015-9484-7}
}

\strong{Data used in:}
\itemize{
\item Culpepper, S. A. (2015). Bayesian estimation of the dina model with
gibbs sampling. \emph{Journal of Educational and Behavioral Statistics},
\emph{40}(5), 454–476. \url{https://doi.org/10.3102/1076998615595403}
\item Culpepper, S. A., & Balamuta, J. J. (2017). A Hierarchical Model for
Accuracy and Choice on Standardized Tests. \emph{Psychometrika}, \emph{82}(3),
820–845. \url{https://doi.org/10.1007/s11336-015-9484-7}
}
}
\keyword{datasets}
