% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecpe-data.R
\docType{data}
\name{qmatrix_ecpe}
\alias{qmatrix_ecpe}
\title{Examination for the Certificate of Proficiency in English (ECPE) Expert-Derived Q matrix}
\format{
A \code{matrix} with dimensions J = 28 (items/rows) and K = 3 (skills/columns).
Each entry in the matrix is either 1, if the item uses the skill, or 0, if
the item does not use the skill. The skills identified by this \code{matrix} are:
\itemize{
\item \code{skill1}: Morphosyntactic rules
\item \code{skill2}: Cohesive rules
\item \code{skill3}: Lexical rules
}

The subjects answered the following assessment items:
\itemize{
\item \code{item01}
\item \code{item02}
\item \code{item03}
\item \code{item04}
\item \code{item05}
\item \code{item06}
\item \code{item07}
\item \code{item08}
\item \code{item09}
\item \code{item11}
\item \code{item12}
\item \code{item13}
\item \code{item14}
\item \code{item15}
\item \code{item16}
\item \code{item17}
\item \code{item18}
\item \code{item19}
\item \code{item20}
\item \code{item21}
\item \code{item22}
\item \code{item23}
\item \code{item24}
\item \code{item25}
\item \code{item26}
\item \code{item27}
\item \code{item28}
}
}
\usage{
qmatrix_ecpe
}
\description{
Examination for the Certificate of Proficiency in English (ECPE) Expert-Derived Q matrix
}
\references{
Templin, J. and Hoffman, L. (2013), Obtaining Diagnostic Classification Model
Estimates Using Mplus. Educational Measurement: Issues and Practice, 32: 37-50.
\url{doi:10.1111/emip.12010}
}
\keyword{datasets}
