% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmatrix-oracle-data.R
\docType{data}
\name{qmatrix_oracle}
\alias{qmatrix_oracle}
\alias{qmatrix_oracle_k3_j20}
\alias{qmatrix_oracle_k4_j20}
\alias{qmatrix_oracle_k5_j30}
\title{Oracle Q Matrices}
\format{
A \code{matrix} with varying numbers of traits (\code{K}) and items (\code{J}).

Specifically:
\itemize{
\item \code{qmatrix_oracle_k3_j20}: 20 items and 3 traits.
\item \code{qmatrix_oracle_k4_j20}: 20 items and 4 traits.
\item \code{qmatrix_oracle_k5_j30}: 30 items and 5 traits.
}
}
\description{
Pre-generated identified Q matrices used in simulation studies to verify
recovery.
}
\details{
Each entry in the matrix is either \code{1}, if the item uses the skill, or \code{0}, if
the item does not use the skill.
}
\keyword{datasets}
