% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability-data.R
\docType{data}
\name{qmatrix_probability_part_one}
\alias{qmatrix_probability_part_one}
\title{Elementary Probability Theory Assessment Expert-Derived Q Matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 12 rows and 4 columns.
}
\usage{
qmatrix_probability_part_one
}
\description{
Elementary Probability Theory Assessment Expert-Derived Q Matrix
}
\details{
Each entry in the matrix is either \code{1}, if the Item uses the Trait, or \code{0}, if
the Item does not use the Trait. The traits identified by this \code{matrix} are:
\itemize{
\item \code{Trait1}
\item \code{Trait2}
\item \code{Trait3}
\item \code{Trait4}
}

For a detailed overview of items, please see \code{\link{items_probability_part_one}}.
}
\references{
\strong{Data originated from:}
\itemize{
\item Heller, J., & Wickelmaier, F. (2013).
Minimum discrepancy estimation in probabilistic knowledge structures.
Electronic Notes in Discrete Mathematics, 42, 49–56.
\doi{10.1016/j.endm.2013.05.145}
}

\strong{Data used in:}
\itemize{
\item Chen, Yinghan, Liu, Y., Culpepper, S. A., & Chen, Y. (2021).
Inferring the number of attributes for the exploratory DINA model.
Psychometrika, 86(1), 30–64. \doi{10.1007/s11336-021-09750-9}
}
}
\keyword{datasets}
