% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy-oracle-data.R
\docType{data}
\name{strategy_oracle}
\alias{strategy_oracle}
\alias{strategy_oracle_k3_j20_s2}
\alias{strategy_oracle_k3_j30_s2}
\alias{strategy_oracle_k3_j40_s2}
\alias{strategy_oracle_k3_j50_s2}
\alias{strategy_oracle_k4_j20_s2}
\alias{strategy_oracle_k4_j30_s2}
\alias{strategy_oracle_k4_j40_s2}
\alias{strategy_oracle_k4_j50_s2}
\title{Strategy Oracle Sets}
\format{
An \code{array} with varying numbers of items (\code{J}), traits (\code{K}), and strategies (\code{S}).

Specifically:
\itemize{
\item \code{strategy_oracle_k3_j20_s2}: 20 items, 3 traits, and 2 strategies.
\item \code{strategy_oracle_k3_j30_s2}: 30 items, 3 traits, and 2 strategies.
\item \code{strategy_oracle_k3_j40_s2}: 40 items, 3 traits, and 2 strategies.
\item \code{strategy_oracle_k3_j50_s2}: 50 items, 3 traits, and 2 strategies.
\item \code{strategy_oracle_k4_j20_s2}: 20 items, 4 traits, and 2 strategies.
\item \code{strategy_oracle_k4_j30_s2}: 30 items, 4 traits, and 2 strategies.
\item \code{strategy_oracle_k4_j40_s2}: 40 items, 4 traits, and 2 strategies.
\item \code{strategy_oracle_k4_j50_s2}: 50 items, 4 traits, and 2 strategies.
}
}
\description{
Pre-generated strategy matrices used in simulation studies to verify
recovery.
}
\details{
Each entry in a strategy is denoted by either \code{1}, if the item uses the skill
under strategy \code{s}, or \code{0}, if the item does not use the skill under strategy
\code{s}.

\strong{Note:} Each matrix in the strategy was generated independently
under the criterion for a strictly identifiable Q matrix.
}
\keyword{datasets}
