% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdmdata-package.R
\docType{package}
\name{ecdmdata-package}
\alias{ecdmdata}
\alias{ecdmdata-package}
\title{ecdmdata: Data Sets for Psychometric Modeling}
\description{
Collection of data sets from assessments that can be used to
evaluate psychometric models.
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} [copyright holder]

Authors:
\itemize{
  \item Steven Culpepper \email{sculpepp@illinois.edu} [copyright holder]
  \item Jeffrey Douglas \email{jeffdoug@illinois.edu} [copyright holder]
}

}
\keyword{internal}
