% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecpe-data.R
\docType{data}
\name{items_ecpe}
\alias{items_ecpe}
\title{Examination for the Certificate of Proficiency in English (ECPE) Item Responses}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2922 rows and 28 columns.
}
\usage{
items_ecpe
}
\description{
Examination for the Certificate of Proficiency in English (ECPE) Item Responses
}
\details{
The subjects answered the following assessment items:
\itemize{
\item \code{item01}
\item \code{item02}
\item \code{item03}
\item \code{item04}
\item \code{item05}
\item \code{item06}
\item \code{item07}
\item \code{item08}
\item \code{item09}
\item \code{item11}
\item \code{item12}
\item \code{item13}
\item \code{item14}
\item \code{item15}
\item \code{item16}
\item \code{item17}
\item \code{item18}
\item \code{item19}
\item \code{item20}
\item \code{item21}
\item \code{item22}
\item \code{item23}
\item \code{item24}
\item \code{item25}
\item \code{item26}
\item \code{item27}
\item \code{item28}
}
}
\references{
\strong{Data originated from:}
\itemize{
\item Templin, J., & Bradshaw, L. (2014). Hierarchical diagnostic
classification models: A family of models for estimating and testing
attribute hierarchies. \emph{Psychometrika}, \emph{79}(2), 317–339.
\doi{10.1007/s11336-013-9362-0}
\item Templin, J., & Hoffman, L. (2013). Obtaining diagnostic classification
model estimates using mplus. \emph{Educational Measurement: Issues and
Practice}, \emph{32}(2), 37–50. \doi{10.1111/emip.12010}
}

\strong{Data used in:}
\itemize{
\item Culpepper, S. A., & Chen, Y. (2019). Development and application of an
exploratory reduced reparameterized unified model. \emph{Journal of
Educational and Behavioral Statistics}, \emph{44}(1), 3–24.
\doi{10.3102/1076998618791306}
}
}
\keyword{datasets}
