% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecpe-data.R
\docType{data}
\name{items_ecpe}
\alias{items_ecpe}
\title{Examination for the Certificate of Proficiency in English (ECPE) Item Responses}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2922 rows and 28 columns.
}
\usage{
items_ecpe
}
\description{
Examination for the Certificate of Proficiency in English (ECPE) Item Responses
}
\details{
The subjects answered the following assessment items:
\itemize{
\item \code{item01}
\item \code{item02}
\item \code{item03}
\item \code{item04}
\item \code{item05}
\item \code{item06}
\item \code{item07}
\item \code{item08}
\item \code{item09}
\item \code{item11}
\item \code{item12}
\item \code{item13}
\item \code{item14}
\item \code{item15}
\item \code{item16}
\item \code{item17}
\item \code{item18}
\item \code{item19}
\item \code{item20}
\item \code{item21}
\item \code{item22}
\item \code{item23}
\item \code{item24}
\item \code{item25}
\item \code{item26}
\item \code{item27}
\item \code{item28}
}
}
\references{
\strong{Data originated from:}
\itemize{
\item Tatsuoka, C. (2002). Data analytic methods for latent partially ordered
classification models. \emph{Journal of the Royal Statistical Society: Series
C (Applied Statistics)}, \emph{51}(3), 337–350.
\url{https://doi.org/10.1111/1467-9876.00272}
\item Tatsuoka, K. K. (1984), Analysis of errors in fraction addition and subtraction
problems (Final Report for Grant No. NIE-G-81-0002). Urbana: University of
Illinois, Computer-Based Education Research Laboratory (CERL).
}

\strong{Data used in:}
\itemize{
\item Chen, Y., Culpepper, S. A., & Liang, F. (2020). A sparse latent class model
for cognitive diagnosis. \emph{Psychometrika}, 1–33.
\url{https://doi.org/10.1007/s11336-019-09693-2}
\item Culpepper, S. A. (2019). Estimating the cognitive diagnosis \eqn{Q} matrix
with expert knowledge: Application to the fraction-subtraction dataset.
\emph{Psychometrika}, \emph{84}(2), 333–357.
\url{https://doi.org/10.1007/s11336-018-9643-8}
\item Culpepper, S. A., & Chen, Y. (2019). Development and application of an
exploratory reduced reparameterized unified model. \emph{Journal of
Educational and Behavioral Statistics}, \emph{44}(1), 3–24.
\url{https://doi.org/10.3102/1076998618791306}
\item Chen, Y., Culpepper, S. A., Chen, Y., & Douglas, J. (2018). Bayesian
estimation of the dina q matrix. \emph{Psychometrika}, \emph{83}(1), 89–108.
\url{https://doi.org/10.1007/s11336-017-9579-4}
}
}
\keyword{datasets}
