% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraction-data.R
\docType{data}
\name{items_fractions}
\alias{items_fractions}
\title{Fraction Subtraction and Addition Assessment Item Responses}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 536 rows and 20 columns.
}
\usage{
items_fractions
}
\description{
Fraction Subtraction and Addition Assessment Item Responses
}
\details{
The subjects answered the following assessment items:
\itemize{
\item \code{Item01}: \eqn{\frac{5}{3}-\frac{3}{4}}{5/3 - 3/4}
\item \code{Item02}: \eqn{\frac{3}{4}-\frac{3}{8}}{3/4 - 3/8}
\item \code{Item03}: \eqn{\frac{5}{6}-\frac{1}{9}}{5/6 - 1/9}
\item \code{Item04}: \eqn{3\frac{1}{2}-2\frac{3}{2}}{3*1/2 - 2*3/2}
\item \code{Item05}: \eqn{4\frac{3}{5}-3\frac{4}{10}}{4*3/5 - 3*4/10}
\item \code{Item06}: \eqn{\frac{6}{7}-\frac{4}{7}}{6/7 - 4/7}
\item \code{Item07}: \eqn{3-2\frac{1}{5}}{3 - 2*1/5}
\item \code{Item08}: \eqn{\frac{2}{3}-\frac{2}{3}}{2/3 - 2/3}
\item \code{Item09}: \eqn{3\frac{7}{8}-2}{3*7/8 - 2}
\item \code{Item10}: \eqn{4\frac{4}{12}-2\frac{7}{12}}{4*4/12 - 2*7/12}
\item \code{Item11}: \eqn{4\frac{1}{3}-2\frac{4}{3}}{4*1/3 - 2*4/3}
\item \code{Item12}: \eqn{\frac{11}{8}-\frac{1}{8}}{11/8 - 1/8}
\item \code{Item13}: \eqn{3\frac{3}{8}-2\frac{5}{6}}{3*3/8 - 2*5/6}
\item \code{Item14}: \eqn{3\frac{4}{5}-3\frac{2}{5}}{3*4/5 - 3*2/5}
\item \code{Item15}: \eqn{2-\frac{1}{3}}{2 - 1/3}
\item \code{Item16}: \eqn{4\frac{5}{7}-1\frac{4}{7}}{4*5/7 - 1*4/7}
\item \code{Item17}: \eqn{7\frac{3}{5}-2\frac{4}{5}}{7*3/5 - 2*4/5}
\item \code{Item18}: \eqn{4\frac{1}{10}-2\frac{8}{10}}{4*1/10 - 2*8/10}
\item \code{Item19}: \eqn{4-1\frac{4}{3}}{4 - 1*4/3}
\item \code{Item20}: \eqn{4\frac{1}{3}-1\frac{5}{3}}{4*1/3 - 1*5/3}
}
}
\references{
\strong{Data originated from:}
\itemize{
\item Tatsuoka, C. (2002). Data analytic methods for latent partially ordered
classification models. \emph{Journal of the Royal Statistical Society: Series
C (Applied Statistics)}, \emph{51}(3), 337–350.
\doi{10.1111/1467-9876.00272}
\item Tatsuoka, K. K. (1984), Analysis of errors in fraction addition and subtraction
problems (Final Report for Grant No. NIE-G-81-0002). Urbana: University of
Illinois, Computer-Based Education Research Laboratory (CERL).
}

\strong{Data used in:}
\itemize{
\item Chen, Y., Culpepper, S. A., & Liang, F. (2020). A sparse latent class model
for cognitive diagnosis. \emph{Psychometrika}, 1–33.
\doi{10.1007/s11336-019-09693-2}
\item Culpepper, S. A. (2019). Estimating the cognitive diagnosis \eqn{Q} matrix
with expert knowledge: Application to the fraction-subtraction dataset.
\emph{Psychometrika}, \emph{84}(2), 333–357.
\doi{10.1007/s11336-018-9643-8}
\item Culpepper, S. A., & Chen, Y. (2019). Development and application of an
exploratory reduced reparameterized unified model. \emph{Journal of
Educational and Behavioral Statistics}, \emph{44}(1), 3–24.
\doi{10.3102/1076998618791306}
\item Chen, Y., Culpepper, S. A., Chen, Y., & Douglas, J. (2018). Bayesian
estimation of the dina q matrix. \emph{Psychometrika}, \emph{83}(1), 89–108.
\doi{10.1007/s11336-017-9579-4}
}
}
\keyword{datasets}
