% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_choice_data.R
\docType{data}
\name{items_spatial}
\alias{items_spatial}
\title{Item Matrix Results of Revised PSVT:R}
\format{A \code{matrix} with 516 observations on the following 30 variables.
\itemize{
\item \code{Item01}: Subject's Response to Item 1.
\item \code{Item02}: Subject's Response to Item 2.
\item \code{Item03}: Subject's Response to Item 3.
\item \code{Item04}: Subject's Response to Item 4.
\item \code{Item05}: Subject's Response to Item 5.
\item \code{Item06}: Subject's Response to Item 6.
\item \code{Item07}: Subject's Response to Item 7.
\item \code{Item08}: Subject's Response to Item 8.
\item \code{Item09}: Subject's Response to Item 9.
\item \code{Item10}: Subject's Response to Item 10.
\item \code{Item11}: Subject's Response to Item 11.
\item \code{Item12}: Subject's Response to Item 12.
\item \code{Item13}: Subject's Response to Item 13.
\item \code{Item14}: Subject's Response to Item 14.
\item \code{Item15}: Subject's Response to Item 15.
\item \code{Item16}: Subject's Response to Item 16.
\item \code{Item17}: Subject's Response to Item 17.
\item \code{Item18}: Subject's Response to Item 18.
\item \code{Item19}: Subject's Response to Item 19.
\item \code{Item20}: Subject's Response to Item 20.
\item \code{Item21}: Subject's Response to Item 21.
\item \code{Item22}: Subject's Response to Item 22.
\item \code{Item23}: Subject's Response to Item 23.
\item \code{Item24}: Subject's Response to Item 24.
\item \code{Item25}: Subject's Response to Item 25.
\item \code{Item26}: Subject's Response to Item 26.
\item \code{Item27}: Subject's Response to Item 27.
\item \code{Item28}: Subject's Response to Item 28.
\item \code{Item29}: Subject's Response to Item 29.
\item \code{Item30}: Subject's Response to Item 30.
}}
\source{
Choice38 Experiment at UIUC during Spring 2014 - Spring 2015
}
\usage{
items_spatial
}
\description{
This data set contains the subject's responses to items. Correct answers are
denoted by 1 and incorrect answers are denoted by 0.
}
\author{
Steven Culpepper and James Balamuta
}
\keyword{datasets}
