% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecpe-data.R
\docType{data}
\name{qmatrix_ecpe}
\alias{qmatrix_ecpe}
\title{Examination for the Certificate of Proficiency in English (ECPE) Expert-Derived Q matrix}
\format{
An object of class \code{q_matrix} (inherits from \code{matrix}) with 28 rows and 3 columns.
}
\usage{
qmatrix_ecpe
}
\description{
Examination for the Certificate of Proficiency in English (ECPE) Expert-Derived Q matrix
}
\details{
Each entry in the matrix is either \code{1}, if the item uses the skill, or \code{0}, if
the item does not use the skill. The skills identified by this \code{matrix} are:
\itemize{
\item \code{skill1}: Morphosyntactic rules
\item \code{skill2}: Cohesive rules
\item \code{skill3}: Lexical rules
}

The subjects answered the following assessment items:
\itemize{
\item \code{item01}
\item \code{item02}
\item \code{item03}
\item \code{item04}
\item \code{item05}
\item \code{item06}
\item \code{item07}
\item \code{item08}
\item \code{item09}
\item \code{item11}
\item \code{item12}
\item \code{item13}
\item \code{item14}
\item \code{item15}
\item \code{item16}
\item \code{item17}
\item \code{item18}
\item \code{item19}
\item \code{item20}
\item \code{item21}
\item \code{item22}
\item \code{item23}
\item \code{item24}
\item \code{item25}
\item \code{item26}
\item \code{item27}
\item \code{item28}
}
}
\references{
\strong{Data originated from:}
\itemize{
\item Templin, J., & Bradshaw, L. (2014). Hierarchical diagnostic
classification models: A family of models for estimating and testing
attribute hierarchies. \emph{Psychometrika}, \emph{79}(2), 317–339.
\url{https://doi.org/10.1007/s11336-013-9362-0}
\item Templin, J., & Hoffman, L. (2013). Obtaining diagnostic classification
model estimates using mplus. \emph{Educational Measurement: Issues and
Practice}, \emph{32}(2), 37–50. \url{https://doi.org/10.1111/emip.12010}
}

\strong{Data used in:}
\itemize{
\item Culpepper, S. A., & Chen, Y. (2019). Development and application of an
exploratory reduced reparameterized unified model. \emph{Journal of
Educational and Behavioral Statistics}, \emph{44}(1), 3–24.
\url{https://doi.org/10.3102/1076998618791306}
}
}
\keyword{datasets}
